First commit
This is the first commit of the project, it contans the structure and an
initial implementation of the game.
It is (obviously) in a very early stage, it has been implemented the
logic to create tetraminos and what in flappy birds are 'pipes' (that
here are wall with a tetramino-shaped socket). There is a basic event
handling (player can move with up/down arrows or mouse, N key creates a
new wall and ESC reset game state).