Implementation of game logic
The game starts to get playable, now the player can be rotated and if it
fits the socket in the wall a simple score gets incremented.
If the wall approaches the player that do not 'enter' in its hole, the
wall 'pushes' the player on the side, and when it reaches the limit it
is a 'game over' (actually the game simply restarts).