Handle touch control and improve game speed
Handle touch control. Touches on the right side of the screen control
the player position, while touches on the left side its rotation.
Improved game speed, now the velocity of elements is proportional to
game's velocity (its FPS), so that the game does not became easier or
harder depending on the device. In particular, the speed is influenced
by the screen width and the game frames per seconds, in a way that every
wall crosses the screen in a fixed amount of time (that also depend on
the current score).